predict_3d_rollball_objects
