+++ README.md
@@ -9,6 +9,7 @@ Just `ai vs ai`
 | Challenge Type      | Prompt                                                                                 | Demo                                                         |
 |---------------------|----------------------------------------------------------------------------------------|--------------------------------------------------------------|
 | 3d_rollball_animals | Use the arrows to rotate the animal to face in the direction of the hand               | [3d_rollball_animals_demo](demo/3d_rollball_animals_demo.py) |
+| 3d_rollball_objects | Use the arrows to rotate the object to face in the direction of the hand               | [3d_rollball_objects_demo](demo/3d_rollball_objects_demo.py) |
 | numericalmatch      | Use the arrows to change the number of objects until it matches the left image         | [numericalmatch_demo](demo/numericalmatch_demo.py)           |
 | hopscotch_highsec   | Use the arrows to move the person to the icon indicated by the colored circle          | [hopscotch_highsec_demo](demo/hopscotch_highsec_demo.py)     |
 | icon_connect        | Using the arrows, connect the same two icons with the dotted line as shown on the left | 🚧                                                           |
