+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "funcaptcha-challenger"
-version = "0.7.0"
+version = "0.8.0"
 description = ""
 authors = ["Madoka <s@madokax.com>"]
 readme = "README.md"
